Made jump consistent. Fixed gold note music.
Jumps need to be the same height, so you need to zero out the vertical velocity before adding upward force to the rigidbody, otherwise if the player is falling, then it will already have a downward velocity that will go against the upward velocity of a jump.
Gold note was muting the guitar slightly at the end, bc we were exiting the rainbow trigger, instead of just muting if we exit the trigger of a note.